requirements.txt
matplotlib~=3.5.3
numpy~=1.23.5
tensorflow~=2.11.0
Pillow~=9.3.0
pandas~=1.5.2
requests~=2.28.1
scikit-learn~=1.1.3